+++ paper/paper.bib
@@ -14,7 +14,7 @@
     volume = {29},
     number = {13},
     pages = {2044-2078},
-    doi = {https://doi.org/10.1002/jcc.21057},
+    doi = {10.1002/jcc.21057},
     year = {2008}
 }
 
@@ -60,17 +60,21 @@
     volume = {275},
     pages = {111-118},
     year = {2016},
-    doi = {https://doi.org/10.1016/j.cattod.2015.10.009},
+    doi = {10.1016/j.cattod.2015.10.009},
     author = {Ivo A.W. Filot and Farid Fariduddin and Robin J.P. Broos and Bart Zijlstra and Emiel J.M. Hensen},
 }
 
-@InCollection{ParaView,
-  author    = {James Ahrens and Berk Geveci and Charles Law},
-  booktitle = {Visualization Handbook},
-  publisher = {Elesvier},
-  title     = {{ParaView}: An End-User Tool for Large Data Visualization},
-  year      = {2005},
-  note      = {{ISBN}~978-0123875822},
+@incollection{ParaView,
+    title = {ParaView: An End-User Tool for Large-Data Visualization},
+    editor = {Charles D. Hansen and Chris R. Johnson},
+    booktitle = {Visualization Handbook},
+    publisher = {Butterworth-Heinemann},
+    address = {Burlington},
+    pages = {717-731},
+    year = {2005},
+    isbn = {978-0-12-387582-2},
+    doi = {10.1016/B978-012387582-2/50038-1},
+    author    = {James Ahrens and Berk Geveci and Charles Law},
 }
 
 @book{OpenDX,
@@ -109,11 +113,17 @@
     year = {2010}
 }
 
-@misc{BoostLibrary,
-    author = {Boost},
-    year   = {2015},
-    title  = {Boost C++ Libraries},
-    howpublished = "{http://www.boost.org/}",
+@Inbook{BoostLibrary,
+    author="Koranne, Sandeep",
+    title="Boost C++ Libraries",
+    bookTitle="Handbook of Open Source Tools",
+    year="2011",
+    publisher="Springer US",
+    address="Boston, MA",
+    pages="127--143",
+    isbn = "978-1-4419-7719-9",
+    doi= " 10.1007/978-1-4419-7719-9_6",
+    url = "https://doi.org/10.1007/978-1-4419-7719-9_6"
 }
 
 @misc{CairoLibrary,
@@ -136,7 +146,10 @@
   journal={Ussr Computational Mathematics and Mathematical Physics},
   year={1976},
   volume={16},
-  pages={10-24}
+  number = {2},
+  pages={10-24},
+  issn = {0041-5553},
+  doi = {10.1016/0041-5553(76)90100-2},
 }
 
 @Article{hunter:2007,
