+++ README.md
@@ -1,7 +1,9 @@
 # EDP – a program for projecting electron densities
 
-[![C/C++ CI](https://github.com/ifilot/edp/actions/workflows/c-cpp.yml/badge.svg)](https://github.com/ifilot/edp/actions/workflows/c-cpp.yml)
+[![Build](https://github.com/ifilot/edp/actions/workflows/build.yml/badge.svg)](https://github.com/ifilot/edp/actions/workflows/build.yml)
+[![Documentation](https://github.com/ifilot/edp/actions/workflows/docs.yml/badge.svg)](https://edp.imc-tue.nl)
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
+[![status](https://joss.theoj.org/papers/5544210e68408b1f00a6fb802b7745e8/status.svg)](https://joss.theoj.org/papers/5544210e68408b1f00a6fb802b7745e8)
 
 * Detailed documentation can be found at https://edp.imc-tue.nl.
 * See [the contributing guide](CONTRIBUTING.md) for detailed instructions how to make contributions.
