+++ README.md
@@ -1,5 +1,6 @@
 # EDP – a program for projecting electron densities
 
+![GitHub tag (latest SemVer)](https://img.shields.io/github/v/tag/ifilot/edp?label=version)
 [![Build](https://github.com/ifilot/edp/actions/workflows/build.yml/badge.svg)](https://github.com/ifilot/edp/actions/workflows/build.yml)
 [![Documentation](https://github.com/ifilot/edp/actions/workflows/docs.yml/badge.svg)](https://edp.imc-tue.nl)
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
