### 2023-04-28 #5417 @editorialbot
[REVIEW]: EDP: a program for projecting electron densities from VASP onto planes
**Submitting author:** @ifilot (<a href="http://orcid.org/0000-0003-1403-8379">Ivo Filot</a>)
**Repository:** https://github.com/ifilot/edp
**Branch with paper.md** (empty if default branch):
**Version:** 2.0.3
