### 2023-05-01 #5417 @igor-marques
@ifilot has developed a straightforward and user-friendly tool, accompanied by a concise and well-produced paper. EDP effectively simplifies the selection of planes for electron density representation and enables representation along an axis or around a selected atomic centre.

I have identified only a few minor issues that could be addressed for further clarity:

1. The MatLab citation appears unusual ("Inc., T. M." instead of "The MathWorks Inc.").
2. There is a misformatting at the end of the User Interface - Additional features section: "The sampling points are based on the coordinates of the 23<sup>rd</sup> order Lebedev quadrature."
3. The Python script dependencies should be mentioned in the Examples - Special features section. Although matplotlib.pyplot and numpy are likely available for most potential users of EDP, it is still important to provide this information.
4. I could not find information related to the "Community guidelines" checkpoint. Only in the Publications section is there a suggestion/note for the community to contact the developer to add missing publications to the list.
5. In the Z-averaging use-case, would it be beneficial to include an option for centring the data/plot?

Overall, the tool and accompanying paper are well-executed, and addressing these minor points should enhance their clarity.
