### 2023-05-01 #5417 @ifilot
I thank @igor-marques for his valuable time and effort reviewing the software and offering useful suggestions for further improving the documentation. Very much appreciated.

1. I forgot to use double accolades for the author name in `paper.bib`. This has been corrected and the company name is now correctly shown in the references.
2. Indeed, I used HTML code instead of the reST for the superscript.
3. I have added a note box directly under the [section header](https://edp.imc-tue.nl/examples.html#special-features) to emphasize upon these dependencies.
4. I have added a new section [Community guidelines](https://edp.imc-tue.nl/community_guidelines.html) to the documentation to make it (hopefully) very easy for potential user where to find this information.

@igor-marques : I am sorry to have to ask, but could you please clarify upon item #5? Could you perhaps give an example of what this feature should look like? In its current implementation, the averaging of the electron density is carried out per z-coordinate for all planes whose normal vector points in the z-direction. Given a fixed unit cell, one thus obtains non-zero values in the location where the slab is. If the user has placed the slab at the center of the unit cell, the z-averaging will reflect this. Similarly, if the user has positioned the slab at the origin of the unit cell, there will be non-zero values for the averaging around the bottom and top of the graph. Is your intention that the script should accommodate a feature that the user can 'roll' the values in the array as implemented in [this Python function](https://numpy.org/doc/stable/reference/generated/numpy.roll.html)?
