### 2023-05-01 #5417 @igor-marques
@ifilot thanks for going over my suggestions so quickly!
concerning item #5, this is what is generated:
![al_fcc111_zavg bak](https://user-images.githubusercontent.com/21105728/235527892-ced2ee89-b451-4e81-9e0b-1370806b7e43.jpg)

and this is what I meant would be an interesting option "out-of-the-box" (I just "rolled" the X-axis values of this plot to achieve this, so this is totally possible to do for the average user):
![al_fcc111_zavg](https://user-images.githubusercontent.com/21105728/235528301-7da298d2-8bc7-4cea-841a-5c0b2487cb4c.jpg)

If the user has the system centred, this would not be an issue, I assume.

(sorry for alluding to this, but is a defect from many years looking at electron densities profiles from phospholipid bilayers - eg., Fig 11 in https://pubs.rsc.org/en/content/articlepdf/2019/ra/c8ra08441f )
