+++ src/CMakeLists.txt
@@ -29,11 +29,11 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_SOURCE_DIR}/cmake/modules )
 SET(VERSION_MAJOR "2")
 SET(VERSION_MINOR "0")
 SET(VERSION_MICRO "1")
-configure_file(config.h.in config.h @ONLY)
-if(EXISTS "${CMAKE_BINARY_DIR}/config.h")
-    MESSAGE("[USER] Creating ${CMAKE_SOURCE_DIR}/config.h from ${CMAKE_BINARY_DIR}/config.h.in")
+configure_file(${CMAKE_SOURCE_DIR}/config.h.in ${CMAKE_SOURCE_DIR}/config.h @ONLY)
+if(EXISTS "${CMAKE_SOURCE_DIR}/config.h")
+    MESSAGE("[USER] Creating ${CMAKE_SOURCE_DIR}/config.h from ${CMAKE_SOURCE_DIR}/config.h.in")
 else()
-    MESSAGE("[USER] Could not find configuration file in ${CMAKE_BINARY_DIR}!")
+    MESSAGE("[USER] Could not find configuration file in ${CMAKE_SOURCE_DIR}!")
 endif()
 
 # Enable release build
