### 2023-05-02 #5417 @ifilot
@igor-marques Thank you very much for the clarification and the example. I have adjusted the example script and added the following lines (see also: https://github.com/ifilot/edp/blob/master/examples/al_fcc111_zavg.py#L13)

```
# to center the x-axis such that the origin lies at in the middle
# of the graph, uncomment the line below
#data[:,0] = data[:,0] - (np.max(data[:,0]) - np.min(data[:,0])) / 2
```

I hope this addition adequately addresses item #5.
