### 2023-05-11 e58711d
Fixing bug in matrix transpose
