### 2023-06-12 #5417 @rkurchin
👋 Hi @1mikegrn, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
