### 2023-06-27 #5417 @1mikegrn
@ifilot has built a CLI application for parsing CHGCAR, PARCHG, and LOCPOT files and rendering them out to image files depending on user input, or optionally a binary file for use elsewhere. It takes data stored in txt format and projects it onto 2D and 1D planes for visualization.

In terms of minor revisions, the only thing of note I'll say is that the python requirements should be compiled into a [requirements.txt file](https://packaging.python.org/en/latest/tutorials/installing-packages/#requirements-files), probably in the `scripts` directory. Python users who then want to make use of your provided scripts can then simply execute `python -m pip install -r ./requirements.txt` from within the scripts folder and they should be good to go. This is pretty standard practice.

This next part isn't a revision requirement, but I do want to encourage you to consider if and how you expect this application to change over time, and what tools you plan on using in order to best facilitate you CLI's evolution.

Part of the app is written in C++, which, when I look at this I consider it to be the "core application"; and then there are some Python scripts which feel somewhat like after-the-fact inclusions, and the cohesion between the two ecosystems isn't really there. It's not necessarily a "problem" per se as they exist independently. But at the same time, to make use of the full scope of what you've built here, you have to be reasonably well versed in both paradigms and hop between the two, using C++ to parse and generate a binary file, and then python to parse the binary file to a data representation that can be used in MPL.

If you see yourself in the future making more use of Python's package ecosystem to run your analytics, you may wish to consider refactoring your application by writing your controller in python, and then using its C/Python API to directly build the C++ model, which can then turn around and pass resultant data back to python for viewing.

Now, that being said, in terms of core functionality everything seems to check out, app compiles from source, tests pass, and the CLI seems to properly handle user input as I would expect. Pending that minor revision regarding the requirements file, I'd say it should be accepted.
