### 2023-06-28 9d29e4b
Adding requirements.txt file, see issue #23
