+++ scripts/requirements.txt
@@ -1,0 +1,2 @@
+matplotlib==3.7.1
+numpy==1.17.4
