### 2023-06-28 #5417 @ifilot
@1mikegrn Thank you very much for your time and effort reviewing the manuscript and the program.

I would like to express my gratitude for your recommendation regarding the inclusion of a requirements.txt file within the ./scripts folder. I have implemented this suggestion. Furthermore, I will consider this valuable advice for my other ongoing program development endeavors.

I appreciate your valuable insights regarding the potential challenges posed by the utilization of two programming languages, which may impose a burden on users by requiring them to familiarize themselves with two distinct programming paradigms. Your observation about the Python scripts serving as an add-on (indeed an afterthought) is accurate. Initially, the program was designed with the primary objective of generating isosurfaces solely through the command-line interface (CLI), wherein the C++ program was responsible for producing .png images. From my perspective, the C++ program was considered "complete" in that regard.

However, upon receiving feedback from end-users, a common request emerged, which highlighted their desire for the ability to customize these images. Many users suggested the option of loading the raw data into Python and utilizing matplotlib to build the images, offering more flexibility. Taking this suggestion into consideration, I have implemented the necessary tools to fulfill this requirement. It is important to note that my intention was never to construct a fully comprehensive controller routine around the core C++ application.

Nonetheless, I believe it may be highly beneficial to incorporate such functionality in the next iteration of the program.
