### 2023-06-28 #5417 @ifilot
@rkurchin

I have incorporated the suggestions provided by the reviewers, aiming to ensure that both the manuscript and program meet the required standards for publication in JOSS. However, if there are any outstanding matters or additional improvements you would like me to address, I am more than willing to consider and implement them accordingly.

The initial submission was marked as version 2.0.0. But after incorporating feedback from the reviewer, we are now at version 2.0.3. I just wanted to check if we need to do anything differently during the submission process to account for this updated version number.
