### 2023-06-30 #5417 @rkurchin
Editorial comments:
- I would suggest adding one or two more sentences to the "Summary" section addressing specifically what EDP does in the context of the other information you've provided. Quoting from JOSS paper guidelines, the section should provide:
  > A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- line 26: I think there's an extra space before the comma after "interactive use"
- line 28: hyphenate both "C++-based" and "command-line" as they are serving as [compound adjectives](https://www.grammar-monster.com/lessons/hyphens_in_compound_adjectives.htm)
- 29, 31, 32, 36, and other places: Choose a convention regarding having a space before citations (as you have earlier in the paper) or not and be consistent. I think it looks better with the space.
- 31: add "and" before `TCLAP`
- 32-33: hyphenate "Linux-based"
- 52: I would change "fixed" to "set" here
- 85, 95: fix capitalization of C++
