+++ paper/paper.md
@@ -8,6 +8,7 @@ tags:
 authors:
   - name: I.A.W. Filot
     orcid: 0000-0003-1403-8379
+    corresponding: true
     affiliation: 1
 affiliations:
  - name: Inorganic Materials and Catalysis, Department of Chemical Engineering and Chemistry, Eindhoven University of Technology
@@ -18,25 +19,27 @@ bibliography: paper.bib
 
 # Summary
 
-The electron density is a fundamental concept in electronic structure calculations, as it provides a detailed description of the distribution of electrons in a material or molecule. In many electronic structure methods, the electron density is used as the basic quantity from which other properties such as the energy, potential, and forces can be derived. The electron density provides insight into chemical bonding and reactivity and understanding the electron density (and its redistribution upon chemical bond breaking or formation) is crucial for predicting and interpreting the properties of materials and molecules.[@martin:2004]
+The electron density is a fundamental concept in electronic structure calculations, as it provides a detailed description of the distribution of electrons in a material or molecule. In many electronic structure methods, the electron density is used as the basic quantity from which other properties such as the energy, potential, and forces can be derived. The electron density provides insight into chemical bonding and reactivity and understanding the electron density (and its redistribution upon chemical bond breaking or formation) is crucial for predicting and interpreting the properties of materials and molecules. [@martin:2004]
 
 The electron density is a scalar field, which means that it is a function that assigns a scalar value to each point in space. In the case of the electron density, this scalar value represents the probability density of finding an electron at that point in space. Specialized visualization tools and techniques are often required to effectively visualize scalar fields such as the electron density. These tools may include contour plots, isosurface rendering, and volume rendering, among others. Visualizing the electron density is essential for gaining insights into the electronic properties and behavior of materials and molecules.
 
+`EDP` is a C++-based command-line utility designed to perform the projection of the electron density scalar field onto a plane. Subsequently, this plane is rendered onto a canvas using a color map and stored as a PNG file.
+
 # Statement of need
 
-The visualization of scalar fields using contour plots or heat maps is ubiquitous in science and engineering and plenty of general purpose programs such as `Open Data Explorer` [@OpenDX] and `ParaView` [@ParaView] readily accomodate this task. There exist also programs that specifically cater to `VASP`, such as `Vesta` [@momma:2011] and `sisl` [@papior:2023]. These programs are typically designed for interactive use , `sisl` being a notable exception, and utilize a graphical user interface of some sort. In contrast, `EDP` is a C++ based command line tool that carries out the projection of the electron density scalar field as stored in `VASP`[@hafner:2008] CHGCAR or PARCHG file onto a plane. This plane is then rendered onto a canvas using a color map and stored as a PNG file. `EDP` uses a minimal set of dependencies, i.e. Boost[@BoostLibrary], Cairo[@CairoLibrary], `TCLAP`[@TclapLibrary], Eigen3[@eigenweb], which are all readily available on modern Linux based operating systems.
[…]
