+++ paper/paper.bib
@@ -109,13 +109,13 @@
 @MISC{eigenweb,
     author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
     title = {Eigen v3},
-    howpublished = {http://eigen.tuxfamily.org},
+    url = "http://eigen.tuxfamily.org",
     year = {2010}
 }
 
 @Inbook{BoostLibrary,
     author="Koranne, Sandeep",
-    title="Boost C++ Libraries",
+    title={"Boost {C++} Libraries"},
     bookTitle="Handbook of Open Source Tools",
     year="2011",
     publisher="Springer US",
@@ -130,14 +130,14 @@
     author = {Cairo},
     year   = {2005},
     title  = {Cairo},
-    howpublished = "{https://www.cairographics.org/}",
+    url = "https://www.cairographics.org/",
 }
 
 @misc{TclapLibrary,
     author = {Michael E. Smoot and Daniel Aarno and others},
     year   = {2009},
-    title  = {Templatized C++ Command Line Parser Library},
-    howpublished = "{https://tclap.sourceforge.net/}",
+    title  = {Templatized {C++} Command Line Parser Library},
+    url = "https://tclap.sourceforge.net/",
 }
 
 @article{lebedev:1976,
