### 2023-07-04 dfc6224
Update paper.md
Fixing spelling error.
